requirements.txt
streamlit
pandas
requests
python-dateutil
streamlit-autorefresh
pydeck
plotly
